requirements.txt
biopython==1.81
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cmake==3.27.7
dgl==1.1.2+cu117
fair-esm==2.0.0
filelock==3.12.4
fsspec==2023.9.2
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
lit==17.0.3
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.26.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.2
pandas==1.5.0
Pillow==10.1.0
psutil==5.9.6
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
sympy==1.12
threadpoolctl==3.2.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.66.1
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
